Fixed issues in README.rst
The link to the git file on GitHub was incorrect, as it pointed to github.org instead of github.com. This isn't an issue in current web browsers, but might pose an issue in other places (besides just not being correct).

Otherwise, there seemed to be a slight issue on line 30 of the file. This too was changed.